+++ src/Network/Network.Test/ScenarioTests/NetworkManagerTests.cs
@@ -98,5 +98,13 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-NetworkManagerResourceMinimumParameterCreate");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.nsmdev)]
+        public void NetworkManagerIpamPoolCRUD()
+        {
+            TestRunner.RunTestScript("Test-NetworkManagerIpamPoolCRUD");
+        }
     }
 }